qcacld-3.0: increase the max supported number of MAC ACL
Extend type of the variables for the number of MAC addresses
in ACL, from 255(uint8_t) to 65535(uint16_t).
With this change, it can support up to 65535 MAC addresses in
theory, after modifying the MACRO 'MAX_ACL_MAC_ADDRESS'.

CRs-Fixed: 2856707

+++ core/hdd/src/wlan_hdd_hostapd_wext.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2020 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012-2021 The Linux Foundation. All rights reserved.
  *
  * Permission to use, copy, modify, and/or distribute this software for
  * any purpose with or without fee is hereby granted, provided that the
@@ -288,7 +288,7 @@ static QDF_STATUS hdd_print_acl(struct hdd_adapter *adapter)
 {
 	eSapMacAddrACL acl_mode;
 	struct qdf_mac_addr maclist[MAX_ACL_MAC_ADDRESS];
-	uint8_t listnum;
+	uint16_t listnum;
 	struct sap_context *sap_ctx;
 
 	sap_ctx = WLAN_HDD_GET_SAP_CTX_PTR(adapter);
